pyproject.toml
[tool.poetry]
name = "testchat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvloop = "^0.19.0"
aiohttp = "^3.9.5"
websockets = "^12.0"
pymongo = "^4.8.0"
python-dotenv = "^1.0.1"
motor = "^3.5.0"
redis = "^5.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
